Build fixes for OS X * Replaced usage of boost::locale (in the logger) with standard C++ functions for converting wide strings to multi-byte strings * Fixed a bunch of uses of size_t that really should have been int (since size_t is unsigned) * Only add the admin escalation option for psmoveservice in MSVC

+++ psmoveservice/ControllerManager.h
@@ -4,7 +4,7 @@
 //-- includes -----
 
 //-- constants -----
-static const size_t k_max_psmove_controllers= 5;
+static const int k_max_psmove_controllers= 5;
 
 //-- definitions -----
 class ControllerManager
